requirements.txt
flask
stripe
requests
python-dotenv
gunicorn